[PR - 3주차 진문장] 추가 요구사항 구현

+++ src/main/java/com/hhplus/precourse/post/service/DeletePostService.java
@@ -1,12 +1,15 @@
 package com.hhplus.precourse.post.service;
 
-import com.hhplus.precourse.common.exception.NotFoundException;
-import com.hhplus.precourse.post.repository.PostRepository;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.hhplus.precourse.common.exception.BadRequestException;
+import com.hhplus.precourse.common.exception.NotFoundException;
 import static com.hhplus.precourse.common.support.ApplicationStatus.POST_NOT_FOUND;
+import static com.hhplus.precourse.common.support.ApplicationStatus.UNAUTHORIZED;
+import com.hhplus.precourse.post.repository.PostRepository;
+
+import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
@@ -14,11 +17,14 @@ public class DeletePostService {
     private final PostRepository postRepository;
 
     @Transactional
-    public void delete(long id, String password) {
+    public void delete(long id, long userId) {
         var post = postRepository.findById(id)
             .orElseThrow(() -> new NotFoundException(POST_NOT_FOUND));
 
-        post.validatePassword(password);
+        if (!post.isAuthor(userId)) {
+            throw new BadRequestException(UNAUTHORIZED, "작성자만 삭제할 수 있습니다.");
+        }
+
         postRepository.delete(post);
     }
 }
